Move RNG fake to fixtures folder
