+++ frontend/tsconfig.json
@@ -27,6 +27,7 @@
     "emitDecoratorMetadata": true
   },
   "include": [
-    "src"
+    "src", 
+    "cypress"  
   ]
 }
